pyproject.toml
[project]
name = "griptape-maya"
version = "0.1.0"
description = "Maya tools build with Griptape"
authors = [
    {name = "Jason Schleifer",email = "jason.schleifer@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

python-dotenv = "^1.0.1"
griptape = { version = "^1.3.1", extras = ["drivers-prompt-anthropic", "drivers-rerank-cohere", "drivers-vector-pgvector", "drivers-prompt-ollama"]}
maya-stubs = "^0.4.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
